Int to Float

+++ main.py
@@ -1,5 +1,5 @@
-num_a = int(input("Enter the 1st Number: "))
-num_b = int(input("Enter the 2nd Number: "))
+num_a = float(input("Enter the 1st Number: "))
+num_b = float(input("Enter the 2nd Number: "))
 operation = input("Which operation do you want to perform?(*,+,-,/): ")
 
 if operation == "*":
